% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{getPackagePath}
\alias{getPackagePath}
\title{function to find package folder}
\usage{
getPackagePath(name = "ffIntro", setAsWD = FALSE)
}
\arguments{
\item{char}{, name of package}

\item{booleen}{, set work directory to path of package?}
}
\value{
path of package
}
\description{
function to find package folder
}
\examples{
getPackagePath(name="base",setAsWD=FALSE)

}

